# swagger.json
{
    "info": {
        "version": "V1.0",
        "description": "## Authentication\n\nAuthenticate requests by including an Authorization header.  For example:\n\n```bash\ncurl 'http://35.154.156.45:3000/api/users/myEntityId' -H \"Authorization: Bearer <auth token>\"\n```\n\n## Generate Token\nTo retrieve an auth token, POST a username/password combination to /api/auth/users.  For example:\n\n```bash\ncurl -X POST 'http://35.154.156.45:3000/api/auth/users' -d \"username=USERNAME&password=PASSWORD\"\n```\n\n",
        "title": "Lend Ledger V0.1 API Docs"
    },
    "paths": {
...
        "/payment/lend": {
            "post": {
                "security": [
                    {
                        "oauth2": [
                            "rw:Lend"
                        ]
                    }
                ],
                "summary": "Here you can lend money.",
                "responses": {
                    "201": {
                        "description": "Successfully created",
                        "schema": {
                            "$ref": "#/definitions/Lend"
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    }
                },
                "parameters": [
                    {
                        "schema": {
                            "$ref": "#/definitions/Lend"
                        },
                        "required": true,
                        "description":"Here you can lend the asset. For lending you have to specify receiver's id.",
                        "name": "body",
                        "in": "body"
                    }
                ],
                "tags": [
                    "Payment"
                ]
            }
        },
        "/payment/exchange": {
            "post": {
                "security": [
                    {
                        "oauth2": [
                            "rw:Exchange"
                        ]
                    }
                ],
                "summary": "Exchange",
                "responses": {
                    "201": {
                        "description": "Successfully created",
                        "schema": {
                            "$ref": "#/definitions/Exchange"
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    }
                },
                "parameters": [
                    {
                        "schema": {
                            "$ref": "#/definitions/Exchange"
                        },
                        "required": true,
                        "description":"Here you can exchange the fiat into the asset. For that, you have to mention your acount id, seed,asset type and amount.",
                        "name": "body",
                        "in": "body"
                    }
                ],
                "tags": [
                    "Payment"
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "oauth2": {
            "flow": "authorizationCode",
            "tokenUrl": "/o/token/",
            "type": "oauth2",
            "authorizationUrl": "/docs/oauth2/authorize",
            "scopes": {
                "r:Account": "List assertions in a User's Backpack",
                "r:profile": "See who you are",
                "rw:backpack": "Upload badges into a User's Backpack",
                "rw:badgeuserAdmin": "rw:badgeuserAdmin",
                "rw:issuer": "Create and update Issuers, create and update Badgeclasses, and award Assertions",
                "rw:profile": "Update your own User profile"
            }
        }
    },
    "definitions": {
...
        "Lend": {
            "required": [],
            "properties": {
                "lenderId": {
                    "type": "string",
                    "description": "Please enter your account id",
                    "format": "string"
                },
                "borrowerId": {
                    "type": "string",
                    "description": "Please enter destination account id.",
                    "format": "string"
                },
                "amount": {
                    "type": "string",
                    "description": "Please enter amount.",
                    "format": "string"
                },
                "remarks": {
                    "type": "string",
                    "description": "You can specify remarks.",
                    "format": "string"
                },
                "assetName": {
                    "type": "string",
                    "description": "Choose asset name.",
                    "format": "string"
                },
                "lenderSeed": {
                    "type": "string",
                    "description": "Please enter your seed.",
                    "format": "string"
                }
            }
        },
        "Exchange": {
            "required": [],
            "properties": {
                "borrowerId": {
                    "type": "string",
                    "description": "borrower id",
                    "format": "Please enter your account id."
                },
                "amount": {
                    "type": "string",
                    "description": "amount",
                    "format": "Please enter amount"
                },
                "remarks": {
                    "type": "string",
                    "description": "remarks",
                    "format": "You can specify remarks."
                },
                "assetName": {
                    "type": "string",
                    "description": "Please enter asset name.",
                    "format": "string"
                },
                "borrowerSeed": {
                    "type": "string",
                    "description": "Please enter your seed.",
                    "format": "string"
                }
            }
        },
...
    },
    "security": [
        {
            "oauth2": []
        }
    ],
    "swagger": "2.0"
}